# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ticket Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List tickets
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, pending, on_hold, solved, closed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, normal, high, urgent]
        - name: assigneeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: organizationId
          in: query
          schema:
            type: ["integer", "null"]
        - name: contactId
          in: query
          schema:
            type: ["integer", "null"]
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, updatedAt, priority, dueAt]
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Ticket list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketPage"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Created ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Not found
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    delete:
      operationId: deleteTicket
      tags: [tickets]
      summary: Delete a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tickets/{id}/comments:
    get:
      operationId: listTicketComments
      tags: [comments]
      summary: List comments for a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createTicketComment
      tags: [comments]
      summary: Add a comment to a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  # ─── CONTACTS ─────────────────────────────────────────────────
...
  /contacts/{id}/tickets:
    get:
      operationId: listContactTickets
      tags: [contacts]
      summary: List tickets for a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"

  # ─── ORGANIZATIONS ────────────────────────────────────────────
...
components:
  schemas:
...
    Ticket:
      type: object
      required: [id, subject, status, priority, type, channel, createdAt, updatedAt]
      properties:
        id:
          type: integer
        subject:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, pending, on_hold, solved, closed]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        type:
          type: string
          enum: [question, incident, problem, task]
        channel:
          type: string
          enum: [email, chat, phone, web, api]
        assigneeId:
          type: ["integer", "null"]
        assignee:
          $ref: "#/components/schemas/Agent"
        requesterId:
          type: ["integer", "null"]
        requester:
          $ref: "#/components/schemas/Contact"
        organizationId:
          type: ["integer", "null"]
        organization:
          $ref: "#/components/schemas/Organization"
        tags:
          type: array
          items:
            type: string
        dueAt:
          type: ["string", "null"]
        firstResponseAt:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
        satisfaction:
          type: ["string", "null"]
          enum: [good, bad, null]
        commentCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    TicketInput:
      type: object
      required: [subject, status, priority, type, channel]
      properties:
        subject:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [open, pending, on_hold, solved, closed]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        type:
          type: string
          enum: [question, incident, problem, task]
        channel:
          type: string
          enum: [email, chat, phone, web, api]
        assigneeId:
          type: ["integer", "null"]
        requesterId:
          type: ["integer", "null"]
        organizationId:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        dueAt:
          type: ["string", "null"]
    TicketUpdate:
      type: object
      properties:
        subject:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [open, pending, on_hold, solved, closed]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        type:
          type: string
          enum: [question, incident, problem, task]
        channel:
          type: string
          enum: [email, chat, phone, web, api]
        assigneeId:
          type: ["integer", "null"]
        requesterId:
          type: ["integer", "null"]
        organizationId:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        dueAt:
          type: ["string", "null"]
        satisfaction:
          type: ["string", "null"]
    TicketPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Ticket"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── CONTACT ──────────────────────────────────────────────
    Contact:
      type: object
      required: [id, name, email, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        organizationId:
          type: ["integer", "null"]
        organization:
          $ref: "#/components/schemas/Organization"
        role:
          type: string
          enum: [end_user, agent, admin]
        tags:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        ticketCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Organization:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        domain:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        plan:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        contactCount:
          type: integer
        ticketCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Agent:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [agent, admin, supervisor]
        avatarUrl:
          type: ["string", "null"]
        isOnline:
          type: boolean
        openTicketCount:
          type: integer
        createdAt:
          type: string
...
    Comment:
      type: object
      required: [id, ticketId, body, isPublic, createdAt]
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        body:
          type: string
        isPublic:
          type: boolean
        authorId:
          type: ["integer", "null"]
        authorName:
          type: ["string", "null"]
        authorRole:
          type: ["string", "null"]
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
        isPublic:
          type: boolean
        authorId:
          type: ["integer", "null"]

    # ─── TAG ──────────────────────────────────────────────────
...