# lib/api-zod/src/generated/types/ticketPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ticket Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Ticket } from './ticket';

export interface TicketPage {
  data: Ticket[];
  total: number;
  page: number;
  limit: number;
}